Moved root Vagrantfile into 'scripts' folder to help prevent confusion (running `vagrant up` in the root folder won't work anyway); added 'run' script in the root folder as a proxy to 'vagrant' - for running 'up', 'halt' and 'destroy' for last VM created (not for provisioning - that's done with the setup script); modified curl commands in 'war-deploy.sh' for faster downloading; added 'redeploy.sh' and 'reset-db.sh' helper scripts for working inside the VM and modified other helper scripts.

+++ scripts/rebuild.sh
@@ -3,11 +3,6 @@
 # This script is to be executed from inside the VM to rebuild XNAT.
 # Gradle will be executed and dependencies downloaded in the VM.
 
-echo
-echo Now running the "rebuild.sh" provisioning script.
-
-source vars.sh
-
 OWD=${PWD}
 
 # Exit with error status
@@ -16,40 +11,57 @@ die() {
     exit -1
 }
 
+[[ -e vars.sh  ]] || die "vars.sh file is required for rebuild script"
+
+source vars.sh
+
 echo
 echo Stopping Tomcat...
 sudo service tomcat7 stop
 
 echo
-echo Rebuilding XNAT...
-
-echo
-echo Deleting web app and redeploying XNAT...
+echo Deleting web app and rebuilding XNAT...
 
 sudo rm -Rf /var/lib/tomcat7/webapps/ROOT*
 
-echo
+if [[ -d ${DATA_ROOT}/src/${XNAT_DIR}  ]]; then
 
-if [[ -e /vagrant/${XNAT_SRC##*/} && ${XNAT_SRC##*/} == *.war  ]]; then
-    echo Copying war file...
-    cp -fv /vagrant/${XNAT_SRC##*/} /var/lib/tomcat7/webapps/ROOT.war
-else
-    echo Executing Gradle build
     cd ${DATA_ROOT}/src/${XNAT_DIR}
-    ./gradlew clean war --refresh-dependencies deployToTomcat
+
+    REFRESH=N
+
+    echo
+    read -p "Would you like to refresh dependencies? This takes longer, but may help resolve build issues. [y/N] " REFRESH
+
+    if [[ $REFRESH =~ [Yy] ]]; then
+        echo
+        echo Executing Gradle build with refresh...
+        ./gradlew clean war --refresh-dependencies
+    else
+        echo
+        echo Executing Gradle build...
+        ./gradlew clean war # deployToTomcat # we'll manually copy the war file so we have it if we want to just redeploy
+    fi
+
+    echo
+    cp -v ${DATA_ROOT}/src/${XNAT_DIR}/build/libs/ROOT.war /var/lib/tomcat7/webapps
+
     cd $OWD
+
 fi
 
 # Reset database?
 DESTROY=N
 echo
-read -p "Would you like to empty the database? ALL XNAT DATA IN THIS VM WILL BE DESTROYED. [y/N] " DESTROY
+read -p "Would you like to reset the database? ALL DATABASE DATA IN THIS VM WILL BE DESTROYED. [y/N] " DESTROY
 
 if [[ $DESTROY =~ [Yy] ]]; then
-
+    echo
+    echo "Dropping '${PROJECT}' database."
     dropdb -U $VM_USER $PROJECT
+    echo
+    echo "Creating new '${PROJECT}' database."
     createdb -U $VM_USER $PROJECT
-
 fi
 
 echo
